general tester of single .wav file
This program is intended to be used to verify in greater detail each .wav file that was discarded after "verify_training_keywords".
The .wav file is plotted (also fft) and played after every called function to understand if it was discarded because of a bad recording or if there is a bug in any of the functions.